requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
numpy==2.2.3
pillow==11.1.0
propcache==0.3.0
requests==2.32.3
shapely==2.0.7
urllib3==2.3.0
yarl==1.18.3
